third_party/blink/web_tests/media/media-fragments/TC0032-expected.txt


Title: TC0032
Fragment: 't=3:7'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
